Refactor Services and Controller add authentication

+++ ShareKnowledgeAPI/Implementation/PostService.cs
@@ -1,34 +1,44 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
+using ShareKnowledgeAPI.Authorization;
 using ShareKnowledgeAPI.Database;
 using ShareKnowledgeAPI.Entities;
 using ShareKnowledgeAPI.Exceptions;
 using ShareKnowledgeAPI.Mapper.DTOs;
 using ShareKnowledgeAPI.Seeder;
 using ShareKnowledgeAPI.Services;
+using System.Security.Claims;
 
 namespace ShareKnowledgeAPI.Implementation
 {
     public class PostService : IPostService
     {
+        private readonly IAuthorizationService _authorizationService;
         private readonly IMapper _mapper;
         private readonly ApplicationDbContext _context;
+        private readonly IUserContextService _userContextService;
         private readonly DataSeeder _dataSeeder;
 
-        public PostService(ApplicationDbContext dbContext, IMapper mapper, DataSeeder dataSeeder) 
+        public PostService(ApplicationDbContext dbContext, IMapper mapper, DataSeeder dataSeeder,
+            IAuthorizationService authorizationService, IUserContextService contextService) 
         {
+            _authorizationService = authorizationService;
             _mapper = mapper;
             _dataSeeder = dataSeeder;
             _context = dbContext;
+            _userContextService = contextService;
         }
 
-        public async Task<IEnumerable<PostDto>> GetAllPostsAsync()
+        public async Task<IEnumerable<PostDto>> GetAllPostsAsync(string searchPhrase)
         {
             _dataSeeder.SeedData();
 
             var posts = await _context.Posts
                 .Include(p => p.Comments)
                 .Include(p => p.Categories)
+                .Where(p => searchPhrase == null || (p.Title.ToLower().Contains(searchPhrase.ToLower()) 
+                        || p.Description.ToLower().Contains(searchPhrase.ToLower())))
                 .ToListAsync();
 
             var postDtos = _mapper.Map<IEnumerable<PostDto>>(posts);
@@ -55,7 +65,8 @@ namespace ShareKnowledgeAPI.Implementation
                        
             var post = _mapper.Map<Post>(postDto);
             post.Categories = categories;
-
+            post.CreatedById = _userContextService.GetUserId;
+           
             await _context.Posts.AddAsync(post);
             await _context.SaveChangesAsync();
 
@@ -70,6 +81,14 @@ namespace ShareKnowledgeAPI.Implementation
             if (post is null)
                 throw new NotFoundException($"Post not found");
 
+            var authorizeResult = _authorizationService.AuthorizeAsync(_userContextService.User, post,
+                new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+
+            if (!authorizeResult.Succeeded)
+            {
+                throw new ForbidException("You don't have an access to this resorce.");
+            }
+
             _context.Posts.Remove(post);
             await _context.SaveChangesAsync();
         }
@@ -81,6 +100,14 @@ namespace ShareKnowledgeAPI.Implementation
 
             if (postFromDb is null)
                 throw new NotFoundException($"Post not found");
+
+            var authorizeResult = _authorizationService.AuthorizeAsync(_userContextService.User, postFromDb,
+                new ResourceOperationRequirement(ResourceOperation.Update)).Result;
+
+            if (!authorizeResult.Succeeded) 
+            {
+                throw new ForbidException("You don't have an access to this resorce.");
+            }
 
             postFromDb.Title = post.Title;
             postFromDb.Description = post.Description;
